+++ lib/db/src/schema/index.ts
@@ -2,3 +2,4 @@ export * from "./entries";
 export * from "./giveaways";
 export * from "./referralRewards";
 export * from "./betaSignups";
+export * from "./analyticsEvents";